Generalized code based of off feedback. More in description.
Created object-selector.js to combine ast0.js and ast1.js, could be used to add more objects in the future like powerups or something similar.

Created object-ufo.js and game-box.js to remove the constants.js file and allow us to change the values using the game manager.

Implemented a difficulty component and made the game spawn more objects and increase the velocity and angular velocity of the ufo with time. The increase in speed and number of objects spawned is scaled by the difficulty variable in gamemanager.js

Also renamed objects.js to object-spawner.js as it is a more fitting name.